analysis/requirements_analysis.txt
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jupyter==1.0.0
ipykernel==6.29.4
numpy==1.26.4
